+++ changelog.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Manager::get($manager = 'default')
+- Manager::set(Manager $instance, $manager = 'default')
+
+### Deprecated
+- Manager::instance()
+- Manager::create()
 
 ## [0.9.3] - 2019-10-01
 ### Fixed